requirements.txt
python-telegram-bot==22.6
yfinance
pandas
numpy
